- Removed @inline
