--- jvm/src/test/scala/InlineTest.scala
@@ -1,99 +1,0 @@
-import org.json4s.ast.{JArray, JObject, JNumber, JString}
-import org.scalameter.PerformanceTest.Microbenchmark
-import org.scalameter.api._
-
-object InlineTest extends Microbenchmark {
-  
-  val sizes: Gen[Int] = Gen.range("size")(300000, 1500000, 300000)
-
-  val ints: Gen[Int] = for {
-    size <- sizes
-  } yield size
-
-  val strings: Gen[String] = for {
-    size <- sizes
-  } yield size.toString
-
-  val bigDecimals: Gen[BigDecimal] = for {
-    size <- sizes
-  } yield BigDecimal(size)
-
-  val floats: Gen[Float] = for {
-    size <- sizes
-  } yield size.toFloat
-
-  val maps: Gen[Map[String,Int]] = for {
-    size <- sizes
-  } yield Map(size.toString -> size)
-
-  performance of "Inline" in {
-
-    measure method "A" in {
-      using(strings) in {
-        s => JArray(JString(s))
-      }
-    }
-
-    measure method "B" in {
-      using(bigDecimals) in {
-        b => JArray(JNumber(b))
-      }
-    }
-
-    measure method "C" in {
-      using(bigDecimals) in {
-        b => JArray(JNumber(b)).value.map{
-          case v @ JNumber(_) => v.to[Int]
-        }
-      }
-    }
-
-
-    measure method "D" in {
-      using(floats) in {
-        f => JArray(JNumber(f))
-      }
-    }
-
-    measure method "E" in {
-      using(maps) in {
-        map => JObject(map.map{case (k,v) => (k,JNumber(v))})
-      }
-    }
-  }
-  
-  performance of "NoInline" in {
-    import noInline.org.json4s._
-    measure method "A" in {
-      using(strings) in {
-        s => ast.JArray(ast.JString(s))
-      }
-    }
-    
-    measure method "B" in {
-      using(bigDecimals) in {
-        b => ast.JArray(ast.JNumber(b))
-      }
-    }
-
-    measure method "C" in {
-      using(bigDecimals) in {
-        b => ast.JArray(ast.JNumber(b)).value.map{
-          case v @ ast.JNumber(_) => v.to[Int]
-        }
-      }
-    }
-
-    measure method "D" in {
-      using(floats) in {
-        f => ast.JArray(ast.JNumber(f))
-      }
-    }
-
-    measure method "E" in {
-      using(maps) in {
-        map => ast.JObject(map.map{case (k,v) => (k,ast.JNumber(v))})
-      }
-    }
-  }
-}
