+++ shared/src/main/scala/org/json4s/ast/package.scala
@@ -5,35 +5,35 @@ package object ast {
 
   object BigDecimalConverter {
     implicit val bigDecimalToByte = new BigDecimalConverter[Byte] {
-      @inline def apply(b: BigDecimal): Byte = b.byteValue()
+      def apply(b: BigDecimal): Byte = b.byteValue()
     }
 
     implicit val bigDecimalToInt = new BigDecimalConverter[Int] {
-      @inline def apply(b: BigDecimal): Int = b.intValue()
+      def apply(b: BigDecimal): Int = b.intValue()
     }
 
     implicit val bigDecimalToShort = new BigDecimalConverter[Short] {
-      @inline def apply(b: BigDecimal): Short = b.shortValue()
+      def apply(b: BigDecimal): Short = b.shortValue()
     }
 
     implicit val bigDecimalToLong = new BigDecimalConverter[Long] {
-      @inline def apply(b: BigDecimal): Long = b.longValue()
+      def apply(b: BigDecimal): Long = b.longValue()
     }
 
     implicit val bigDecimalToBigInt = new BigDecimalConverter[BigInt] {
-      @inline def apply(b: BigDecimal): BigInt = b.toBigInt()
+      def apply(b: BigDecimal): BigInt = b.toBigInt()
     }
 
     implicit val bigDecimalToFloat = new BigDecimalConverter[Float] {
-      @inline def apply(b: BigDecimal): Float = b.floatValue()
+      def apply(b: BigDecimal): Float = b.floatValue()
     }
 
     implicit val bigDecimalToDouble = new BigDecimalConverter[Double] {
-      @inline def apply(b: BigDecimal): Double = b.doubleValue()
+      def apply(b: BigDecimal): Double = b.doubleValue()
     }
 
     implicit val bigDecimalToBigDecimal = new BigDecimalConverter[BigDecimal] {
-      @inline def apply(b: BigDecimal): BigDecimal = b
+      def apply(b: BigDecimal): BigDecimal = b
     }
 
   }
